+++ rb_ws/src/buggy/CMakeLists.txt
@@ -36,7 +36,6 @@ install(PROGRAMS
   scripts/watchdog/watchdog_node.py
   scripts/buggy_state_converter.py
   scripts/serial/ros_to_bnyahaj.py
-  scripts/telemetry/telematics.py
   scripts/debug/debug_steer.py
   scripts/vision/detector_node.py
   scripts/debug/set_steer.py
