--- rb_ws/src/buggy/launch/estimator.xml
@@ -1,5 +1,0 @@
-<launch>
-    <node pkg="foxglove_bridge" exec="foxglove_bridge" name="foxglove" namespace="SC"/>
-    <node name="estimator" pkg="buggy" exec="ukf_node.py" output="screen" namespace="SC"/>
-    <node name="telematics" pkg="buggy" exec="telematics.py" namespace="SC" output="screen"/>
-</launch>
