+++ rb_ws/src/buggy/launch/sim_single_vision.xml
@@ -12,10 +12,9 @@
         <param from="$(var config_file)"/>
     </node>
 
-    <node name="telematics" pkg="buggy" exec="telematics.py" namespace="SC" output="screen"/>
 
 
     <node pkg="buggy" exec="watchdog_node.py" name="SC_watchdog" namespace="SC"/>
-    <node pkg="buggy" exec="detector_node.py" name="detector" output="screen" namespace="SC"/>
+    <!-- <node pkg="buggy" exec="detector_node.py" name="detector" output="screen" namespace="SC"/> -->
 
 </launch>
